refactor: consolidate gateway transport hints into shared constants
