+++ gateway/src/http/routes/twilio-sms-webhook.ts
@@ -1,3 +1,4 @@
+import { buildSmsTransportMetadata } from "../../channels/transport-hints.js";
 import type { GatewayConfig } from "../../config.js";
 import { StringDedupCache } from "../../dedup-cache.js";
 import { handleInbound } from "../../handlers/handle-inbound.js";
@@ -50,22 +51,6 @@ function shouldSendRejectionNotice(phoneNumber: string): boolean {
   }
   rejectionNoticeTimestamps.set(phoneNumber, now);
   return true;
-}
-
-export const SMS_CHANNEL_TRANSPORT_HINTS = [
-  "chat-first-medium",
-  "channel-safe-onboarding",
-  "defer-dashboard-only-tasks",
-  "sms-character-limits",
-] as const;
-export const SMS_CHANNEL_TRANSPORT_UX_BRIEF =
-  "SMS is text-only with carrier-imposed message length limits. Keep responses concise and defer rich-media tasks to desktop.";
-
-export function buildSmsTransportMetadata(): { hints: string[]; uxBrief: string } {
-  return {
-    hints: [...SMS_CHANNEL_TRANSPORT_HINTS],
-    uxBrief: SMS_CHANNEL_TRANSPORT_UX_BRIEF,
-  };
 }
 
 function normalizeSmsPayload(
